Incorrect value of lock name in atomic lock manager

+++ lucene/src/test/java/com/orientechnologies/lucene/test/LuceneAutomaticBackupRestoreTest.java
@@ -32,6 +32,8 @@ import com.orientechnologies.orient.server.config.OServerParameterConfiguration;
 import com.orientechnologies.orient.server.handler.OAutomaticBackup;
 import org.junit.*;
 import org.junit.rules.TemporaryFolder;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import java.io.File;
 import java.io.FileInputStream;
@@ -49,21 +51,24 @@ import static org.assertj.core.api.Assertions.assertThat;
 /**
  * Created by Enrico Risa on 07/07/15.
  */
+@RunWith(JUnit4.class)
 public class LuceneAutomaticBackupRestoreTest {
 
   @Rule
-  public TemporaryFolder      tempFolder = new TemporaryFolder();
+  public TemporaryFolder tempFolder = new TemporaryFolder();
 
-  private final static String DBNAME     = "LuceneAutomaticBackupRestoreTest";
-  private String              URL        = null;
-  private String              BACKUPDIR  = null;
-  private String              BACKUFILE  = null;
+  private final static String DBNAME    = "LuceneAutomaticBackupRestoreTest";
+  private              String URL       = null;
+  private              String BACKUPDIR = null;
+  private              String BACKUFILE = null;
 
   private OServer             server;
   private ODatabaseDocumentTx databaseDocumentTx;
 
   @Before
   public void setUp() throws Exception {
+    final String os = System.getProperty("os.name").toLowerCase();
+    Assume.assumeFalse(os.contains("win"));
 
     System.setProperty("ORIENTDB_HOME", tempFolder.getRoot().getAbsolutePath());
 
@@ -109,9 +114,13 @@ public class LuceneAutomaticBackupRestoreTest {
 
   @After
   public void tearDown() throws Exception {
-    dropIfExists();
+    final String os = System.getProperty("os.name").toLowerCase();
 
-    tempFolder.delete();
+    if (!os.contains("win")) {
+      dropIfExists();
+
+      tempFolder.delete();
+    }
 
   }
